Move developer availability and developer dto to dedicated files
