+++ SpreadsheetUtility/SpreadsheetUtility.Library/GanttGenerator.cs
@@ -352,49 +352,4 @@ namespace SpreadsheetUtility.Library
         #endregion
     }
 
-
-    public class DeveloperDto
-    {
-        public string Team { get; set; }
-        public string Name { get; set; }
-        public double DailyWorkHours { get; set; }
-        public string VacationPeriods { get; set; }
-    }
-    public class DeveloperAvailability
-    {
-        public string Name { get; set; }
-        public double DailyWorkHours { get; set; }
-        public List<(DateTime Start, DateTime End)?> VacationPeriods { get; set; }
-
-        public DateTime NextAvailableDateForTasks = DateTime.Today;
-        public List<GanttTask> Tasks { get; set; } = new List<GanttTask>();
-        public double AllocatedHours { get; set; }
-        public double SlackHours { get; set; }
-        public string VacationPeriodsSerialized => string.Join("|", VacationPeriods.Select(v => $"{v?.Start:yyyy-MM-dd};{v?.End:yyyy-MM-dd}"));
-        public DateTime NextAvailableDate(DateTime fromDate)
-        {
-            DateTime date = fromDate > NextAvailableDateForTasks ? fromDate : NextAvailableDateForTasks;
-            while (IsOnVacation(date) || IsWeekend(date))
-            {
-                date = date.AddDays(1);
-            }
-            return date;
-        }
-        private bool IsWeekend(DateTime date)
-        {
-            return date.DayOfWeek == DayOfWeek.Saturday || date.DayOfWeek == DayOfWeek.Sunday;
-        }
-
-        private bool IsOnVacation(DateTime date)
-        {
-            return VacationPeriods.Any(v => v.HasValue && date >= v.Value.Start && date <= v.Value.End);
-        }
-
-        public void SetNextAvailableDate(DateTime date)
-        {
-            NextAvailableDateForTasks = date;
-        }
-    }
-
-
 }
