+++ SpreadsheetUtility/SpreadsheetUtility.Library/Models/DeveloperDto.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SpreadsheetUtility.Library
+{
+    public class DeveloperDto
+    {
+        public string Team { get; set; }
+        public string Name { get; set; }
+        public double DailyWorkHours { get; set; }
+        public string VacationPeriods { get; set; }
+    }
+}
